Created scripts; Corrected syntax on sc files; corrected behavior on tac2asm; created asm2mem assembler

+++ sc/bhaskara.c
@@ -1,7 +1,7 @@
-int a;
-int b;
-int c;
-int bhaskara;
+byte a;
+byte b;
+byte c;
+byte bhaskara;
 void main() {
 	a = 1+2;
 	bhaskara = ((b*2)-(4*a*c))/(2*a);
